requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
cryptography==2.1.3
Django==2.0.6
django-filter==1.1.0
djangorestframework==3.8.2
idna==2.6
Markdown==2.6.11
oci==1.4.3
psycopg2-binary
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.4.0
python-dateutil==2.7.3
pytz==2016.10
requests==2.18.4
six==1.11.0
urllib3==1.22
